[EKYC-4692] Adding authorization validation check

+++ aws/functions/http/delete-webhooks.ts
@@ -7,17 +7,21 @@ import "source-map-support/register";
 export const handler: APIGatewayProxyHandler = Middleware.wrap(async (event, context) => {
 
     context.callbackWaitsForEmptyEventLoop = false;
-    const ctx = Context.getCurrentValue();
+    const tenantAlias = event.pathParameters.alias;
+    console.log(`Admin permission of ${tenantAlias}: `, Context.hasAdminPermission(tenantAlias));
+    if (!Context.hasPlatformAdminPermission() && !Context.hasAdminPermission(tenantAlias)) {
+        throw new Errors.UnauthorizedAccessError();
+    }
     const subscriptionId = event.pathParameters.subscriptionId;
 
     if (!subscriptionId) {
         throw new Errors.InvalidParametersError(`Subscription id not supplied`);
     }
     console.log(subscriptionId);
-    const factory = new ServiceFactory(ctx.identity.tenant);
+    const factory = new ServiceFactory(tenantAlias);
     const repository = await factory.createWebhookEventRepository();
     await repository.deleteWebhook(subscriptionId);
 
-    return Response.ok({result: "OK"});
+    return Response.ok({ result: "OK" });
 
 });
